updated projects and parterns new in sidebar, base news will no longer manage projects/parteners news, they will be managed by ProjectsNews/ParternsNews
